Former-commit-id: d4e3581d70a74b1eda5236650d16b82774d55ebc

+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Commands/Interview/AnswerSingleOptionQuestionCommand.cs
@@ -9,7 +9,7 @@ namespace WB.Core.SharedKernels.DataCollection.Commands.Interview
     {
         public decimal SelectedValue { get; private set; }
 
-        public AnswerSingleOptionQuestionCommand(Guid interviewId, Guid userId, Guid questionId, int[] rosterVector, DateTime answerTime, decimal selectedValue)
+        public AnswerSingleOptionQuestionCommand(Guid interviewId, Guid userId, Guid questionId, decimal[] rosterVector, DateTime answerTime, decimal selectedValue)
             : base(interviewId, userId, questionId, rosterVector, answerTime)
         {
             this.SelectedValue = selectedValue;
